script: Move a bunch of proxy-related code to script_bindings.
